3_APP/requirements.txt
fastapi
uvicorn[standard]
pydantic>=1.10,<3
torch
transformers
streamlit
requests
scikit-learn
folium
streamlit-folium
streamlit-js-eval
psycopg2-binary
sqlalchemy